All scenarios tested (2)
Some bugs fixed (with director scan and disconnection) + building state implemented (but not tested) for PLB

+++ test code/PLB/ESP industry/lib/HAL/IStation.h
@@ -10,7 +10,7 @@ class IStation
 public:
     virtual ~IStation() = default;
     virtual int getId() = 0;
-    virtual uint32_t getDirectorId() = 0;
+    virtual float getDirectorId() = 0;
     virtual void validateDirector(DirectorState directorState) = 0;
     virtual void charge(float power) = 0;
     virtual void switchMode(StationModes mode) = 0;
